% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fingerprint_term}
\alias{fingerprint_term}
\title{Function that fingerprints a single term}
\usage{
fingerprint_term(record, retina_name)
}
\arguments{
\item{record}{single character string containing the term you want to fingerprint}

\item{retina_name}{name of the retina you want to use. See the 'get_retinas' function above}
}
\value{
sfutils Term containing the fingerprinted text
}
\description{
Function that fingerprints a single term
}
