% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{fingerprint_text}
\alias{fingerprint_text}
\title{Function that fingerprints a single text}
\usage{
fingerprint_text(record, retina_name)
}
\arguments{
\item{record}{single character string containing the text you want to fingerprint}

\item{retina_name}{name of the retina you want to use. See the 'get_retinas' function above}
}
\value{
sfutils Document containing the fingerprinted text
}
\description{
Function that fingerprints a single text
}
