% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_retinas}
\alias{get_retinas}
\title{Get the names of the retinas that are currently available}
\usage{
get_retinas()
}
\value{
list containing the names of the retinas that are currently available
}
\description{
Get the names of the retinas that are currently available
}
