% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{retrieve_similar_terms}
\alias{retrieve_similar_terms}
\title{Retrieve similar terms for a fingerprint}
\usage{
retrieve_similar_terms(retina_name, fingerprint, num_terms = 10)
}
\arguments{
\item{retina_name}{name of the retina}

\item{fingerprint}{fingerprint for which you want to retrieve similar terms}

\item{num_terms}{number of similar terms to retrieve}
}
\description{
Retrieve similar terms for a fingerprint
}
