% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{retrieve_terms}
\alias{retrieve_terms}
\title{Retrieve all terms in the retina}
\usage{
retrieve_terms(retina_name)
}
\arguments{
\item{retina_name}{name of the retina}
}
\description{
Retrieve all terms in the retina
}
